--- src/components/UI/Forms/ProductForm/productFormStyles.ts
@@ -1,20 +1,0 @@
-import { Style } from "shared/theme";
-
-export const productFormStyles: Style = {
-  container: {
-    width: "300px",
-    display: "flex",
-    flexDirection: "column",
-    alignItems: "center",
-    m: 5,
-  },
-  field: {
-    backgroundColor: "white",
-    mt: 1,
-  },
-  button: {
-    mt: 1,
-    px: 2,
-    py: 1,
-  },
-};
